Fix small issues

+++ src/engine.h
@@ -274,7 +274,7 @@ struct engine {
 
 #ifdef WITH_LOGGER
   /* Number of particle steps between dumping a chunk of data */
-  int logger_max_steps;
+  short int logger_max_steps;
 
   /* File name of the dump file */
   struct dump *logger_dump;
